Word Report Generator Working Prototype
This took way too long to finish(?) up but now we have reference code to work off when finding file properties and how to format the list of files in a way that can then be inserted into a word doc.

Also Test_FindRenamedFiles.py is just where I'm gonna start writing the renamed file finder script. Stay tuned for that.