Establishment of all working prototypes
contains working scripts for
1. the hidden file finder
2. the renamed file finder
3. the report generator
4. a slightly updated document template

further care will be needed to OOPify the scripts and make them compatible with the rest of the codebase. the report generator and template document will need input from the rest of the team for its completion due to its reliance on imported file lists to display as per the example found in report_generator.py and how it uses hidden_files.py to display required metadata over the template word document.

personal backlog also includes getting the renamed file finder output directed to the report generator script. also getting both finder scripts to read images using filesystem.py + the renamed file output table setup in the template doc. lots of work still needs to be done but this should be a solid checkpoint for now.